move auth and handlers to controllers
