+++ server/server.js
@@ -9,13 +9,10 @@ const bodyParser = require('body-parser');
 const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 const app = express();
-const requestHandlers = require('./request-handlers');
+const requestHandlers = require('./controllers/request-handlers');
 const path = require('path');
 const port = process.env.PORT || 3000;
-const signup = require('./authroutes.js').signup;
-const manualSignIn = require('./authroutes.js').manualSignIn;
-const manualLogout = require('./authroutes.js').manualLogout;
-const isAuthenticated = require('./authroutes.js').isAuthenticated;
+const isAuthenticated = require('./controllers/authroutes.js').isAuthenticated;
 const axios = require('axios')
 const EBAYKEY = process.env.EBAY_KEY;
 const moment = require('moment')
